pyproject.toml
[tool.poetry]
name = "mender"
version = "0.1.0"
description = "Backend for Mender APP"
license = "MIT"
authors = ["Andres Jalife <andres.jalife@mercadolibre.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
alembic = "^1.13.1"
pytest = "^8.2.0"
psycopg2-binary = "^2.9.9"
firebase-admin = "^6.5.0"
pyrebase4 = "^4.8.0"
cryptography = "^42.0.7"
apscheduler = "^3.10.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
